% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesGG.R
\name{bayesGG}
\alias{bayesGG}
\title{Generalized Gompertz distribution}
\usage{
bayesGG(linear_pred, data, J, q, burn, jump, guess)
}
\arguments{
\item{linear_pred}{the linear predictor formula to be included in the model,
in the LHS one put the name of variable to considered as the response
variable and in the RHS one puts the name of explanatory variables,
where all variables are included in data.}

\item{data}{data to be calculated the quantile of interest. It must have a
column with d value containing a censoring indicator, where 1 corresponds
to a non-censored observation and 0 otherwise. It must also contain a x
column with all the regressors.}

\item{J}{number of regressors in the model.}

\item{q}{quantile of interest.}

\item{burn}{number of iterations to be considered as the burn-in.}

\item{jump}{thinning number for the MCMC.}

\item{guess}{initial values for the MCMC algorithm.}
}
\value{
the likelihood function value given the provided information.
}
\description{
This function returns the fit for a Bayesian regression model.
}
