% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likGG.R
\name{likGG}
\alias{likGG}
\title{Generalized Gompertz distribution}
\usage{
likGG(par, q, data, linear_pred, J, log = FALSE)
}
\arguments{
\item{par}{vector with dimension equal to the}

\item{q}{quantile of interest.}

\item{data}{data to be calculated the quantile of interest. It must have a
column with d value containing a censoring indicator, where 1 corresponds
to a non-censored observation and 0 otherwise. It must also contain a x
column with}

\item{linear_pred}{the linear predictor formula to be included in the model,
in the LHS one put the name of variable to considered as the response
variable and in the RHS one puts the name of explanatory variables,
where all variables are included in data.}

\item{J}{number of regressors in the model.}

\item{log}{if the log-likelihood is to be calculated or not. Default is
FALSE.}
}
\value{
the likelihood function value given the provided information.
}
\description{
This function returns the likelihood function value given data, the parameters
 alpha, lambda and mu and q.
}
