% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p0_dGG.R
\name{p0_dGG}
\alias{p0_dGG}
\title{Generalized Gompertz distribution}
\usage{
p0_dGG(alpha, lambda, theta)
}
\arguments{
\item{alpha}{scale parameter. It must be greater than zero and the default
value is -1.}

\item{lambda}{scale parameter. It must be greater than zero and the default
value is 2.}

\item{theta}{shape parameter. It must be greater than zero and the default
 value is 2.
formula}
}
\value{
A
}
\description{
This function returns the cure fraction value given the parameters
 alpha, lambda and theta.
}
