% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGG.R, R/pGG_mm.R
\name{pGG}
\alias{pGG}
\title{Generalized Gompertz distribution}
\usage{
pGG(y, alpha, lambda, mu, q, lower.tail = TRUE, log.p = FALSE)

pGG(y, alpha, lambda, mu, q, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{y}{the value for which one intends to calculate the density function.}

\item{alpha}{scale parameter. It must be greater than zero.}

\item{lambda}{scale parameter. It must be greater than zero.}

\item{mu}{shape parameter. It must be greater than zero.}

\item{q}{quantile of interest.}

\item{lower.tail}{if TRUE returns the lower tail of the cdf function.
Otherwise, returns the probability of being greater than y.}

\item{log.p}{if TRUE returns the log of the cdf. Default if FALSE.}
}
\value{
A

A
}
\description{
This function returns the cdf function value given y, the parameters
 alpha, lambda and mu and q.

This function returns the cdf function value given y, the parameters
 alpha, lambda and mu and q, considering a mixture model.
}
