% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_mu.R
\name{quantile_mu}
\alias{quantile_mu}
\title{Generalized Gompertz distribution}
\usage{
quantile_mu(q = 0.5, alpha = -1, lambda = 2, theta = 2)
}
\arguments{
\item{q}{the value for the quantile to be used in the calculation of the
mu parameter. It must be between 0 and 1 and the default value is 0.5.}

\item{alpha}{scale parameter. It must be greater than zero and the default
value is -1.}

\item{lambda}{scale parameter. It must be greater than zero and the default
value is 2.}

\item{theta}{shape parameter. It must be greater than zero and the default
 value is 2.
formula}
}
\value{
A numeric value for quantile q, given the other parameters.
}
\description{
This function returns the quantile function for the Generalized Gompertz
 distribution given the quantile q, the parameters alpha, lambda and
 theta.
}
