% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_q.R
\name{quantile_q}
\alias{quantile_q}
\title{Generalized Gompertz distribution}
\usage{
quantile_q(mu = 0.5, alpha = -1, lambda = 2, theta = 2)
}
\arguments{
\item{mu}{parameter mu. Defatult value is 0.5
mu parameter. It must be between 0 and 1 and the default value is 0.5.}

\item{alpha}{scale parameter. It must be greater than zero and the default
value is -1.}

\item{lambda}{scale parameter. It must be greater than zero and the default
value is 2.}

\item{theta}{shape parameter. It must be greater than zero and the default
 value is 2.
formula}
}
\value{
A
}
\description{
This function returns the quantile q for the Generalized Gompertz
 distribution given the parameters mu_q, alpha, lambda and theta.
}
