% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesGG.R
\name{bayesGG}
\alias{bayesGG}
\title{Generalized Gompertz distribution}
\usage{
bayesGG(
  linear_pred_mu,
  linear_pred_alpha,
  linear_pred_lambda,
  data,
  q,
  d,
  iter = 1000,
  burn,
  jump,
  guess
)
}
\arguments{
\item{linear_pred_mu}{the linear predictor formula to be included in the 
model for the conditional quantiles; in the LHS one put the name of 
variable to considered as the response variable and in the RHS one puts the
name of explanatory variables, where all variables are included in data.}

\item{linear_pred_alpha}{the formula for the linear predictor of alpha. 
There is no need to put the LHs of the formula}

\item{linear_pred_lambda}{the formula for the linear predictor of alpha. 
There is no need to put the LHs of the formula}

\item{data}{data to be calculated the quantile of interest.}

\item{q}{quantile of interest.}

\item{d}{name of variable containing the censoring indicator in data.}

\item{iter}{number of posterior samples. Default value is 1000.}

\item{burn}{number of iterations to be considered as the burn-in.}

\item{jump}{thinning number for the MCMC.}

\item{guess}{initial values for the MCMC algorithm.}
}
\value{
the likelihood function value given the provided information.
}
\description{
This function returns the fit for a Bayesian regression model.
}
