% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGG_mm.R
\name{dGG_mm}
\alias{dGG_mm}
\title{Generalized Gompertz distribution}
\usage{
dGG_mm(y, alpha, lambda, mu, q, log = FALSE)
}
\arguments{
\item{y}{the value for which one intends to calculate the density function.}

\item{alpha}{scale parameter. It must be greater than zero.}

\item{lambda}{scale parameter. It must be greater than zero.}

\item{mu}{location parameter}

\item{q}{quantile of interest.}

\item{log}{if the log density is to be calculated or not. Default is
FALSE.}
}
\value{
The density value for given y, the parameters alpha, lambda and 
 mu and q.
}
\description{
This function returns the density function value given y, the parameters
 alpha, lambda and mu and q, considering a mixture model.
}
